requirements.txt
annotated-types==0.7.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.18
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-wizard==0.22.3
EditorConfig==0.12.4
Flask==3.0.3
Flask-Caching==2.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib_metadata==8.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
MarkupSafe==2.1.5
more-itertools==10.4.0
nest-asyncio==1.6.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
plotly==5.23.0
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
retrying==1.3.4
ruff==0.4.10
six==1.16.0
SQLAlchemy==2.0.32
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.20.0
Flask-Sessions==0.1.5
gunicorn